requirements.txt
aiohttp==3.6.1
astroid==2.3.1
async-timeout==3.0.1
attrs==19.2.0
certifi==2019.9.11
chardet==3.0.4
idna==2.8
idna-ssl==1.1.0
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
multidict==4.5.2
pylint==2.4.2
requests==2.22.0
six==1.12.0
slackclient==2.2.0
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.6
wrapt==1.11.2
yarl==1.3.0
